+++ src/Braintree/MonetaryAmount.cs
@@ -1,0 +1,13 @@
+namespace Braintree
+{
+    public class MonetaryAmount
+    {
+        public virtual Decimal Value {get; protected set;}
+        public virtual string CurrencyCode {get; protected set;}
+        protected internal MonetaryAmount(Decimal value, string currencyCode)
+        {
+            Value = value;
+            CurrencyCode = currencyCode;
+        }
+    }
+}